+++ changelog.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.16.0.1] - UNRELEASED
+
+- fix gregorianPaschalMoon and gregorianEaster
+- fix negative rollover differences in diffGregorianDurationRollOver and diffJulianDurationRollOver
+
 ## [1.16] - 2026-05-03
 
 - add periodIn